requirements.txt
discord.py
pip
psutil